% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/20065R.R
\name{Gradmat}
\alias{Gradmat}
\title{An implementation of approximately compute the Jacobian Matrix.}
\usage{
Gradmat(parvec, infcn, eps = 1e-06)
}
\arguments{
\item{parvec}{a vector (x1, x2, ..., xk)(vector)}

\item{infcn}{the input function (g1, g2, ..., gm) (vector)}

\item{eps}{a small difference in parvec, with default value 1e-06 (numeric)}
}
\value{
a jacobian matrix (gradient matrix) of function (g1, g2, ..., gm)
}
\description{
This is an implementation of approximately compute the Jacobian
    Matrix of a multivariate function.
}
\examples{
\dontrun{
f=function(x) c(x[1]^3-x[2]-1,x[1]^3-x[2]^2-1)
f(c(1,2))
Gradmat(c(1,2),f)
}
}
