% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/20065R.R
\name{NRroot}
\alias{NRroot}
\title{An implementation of Newton-Raphson Method for root finding for multivariate function.}
\usage{
NRroot(inipar, infcn, nmax = 50, stoptol = 1e-05, eps = 1e-06, gradfunc = NULL)
}
\arguments{
\item{inipar}{an initial value for vector (x1_0, x2_0, ..., xk_0)(vector)}

\item{infcn}{the input function (g1, g2, ..., gm) (vector)}

\item{nmax}{the maximum iteration times, with default 25 (int)}

\item{stoptol}{the tolerance condition for stopping the iteration (numeric)}

\item{eps}{a small difference in parvec, with default value 1e-06 (numeric)}

\item{gradfunc}{the gradient matrix passed from outside, with default NULL (logical)}
}
\value{
a list that contains: the iteration times, the root founded by the method, the initial condition, the function value at root.
}
\description{
An implementation of Newton-Raphson Method for root finding for multivariate function.
}
\examples{
\dontrun{
f=function(x) c(x[1]^3-x[2]-1,x[1]^3-x[2]^2-1)
NRroot(c(1,2),f)
}
}
